+++ mocks/mock-server-data.json
@@ -12,7 +12,7 @@
     ],
     "descriptions": [
       "Комфорное вместительное помещение",
-      "Шикарные аппортоменты в центре города",
+      "Шикарные апартаменты в центре города",
       "Уютное гнёздышко для молодой семьи, курящих просьба не беспокоить",
       "Отличное место с возможностью свободной планировки",
       "Отдельный дом, нуждается в косметическом ремонте. Торг возможен",
@@ -20,7 +20,7 @@
       "Съем от трёх суток",
       "Оплата почасовая, зал находится на цокольном этаже, удобства имеются"
     ],
-    "city": [
+    "cities": [
       "Paris", "Cologne", "Brussels", "Amsterdam", "Hamburg", "Dusseldorf"
     ],
     "previewImages": [
